Moved fixture to separate conftest.py file.
