+++ unit_tests/test_settings.py
@@ -1,31 +1,31 @@
 """Simple unit tests for django-simple-deploy."""
 
-import subprocess
-from time import sleep
+# import subprocess
+# from time import sleep
 from pathlib import Path
 
 import pytest
 
 
-@pytest.fixture(scope='module')
-def tmp_project(tmpdir_factory):
-    """Create a copy of the local sample project, and run simple_deploy
-    against this project. Most tests will examine how the project
-    was modified.
-    """
+# @pytest.fixture(scope='module')
+# def tmp_project(tmpdir_factory):
+#     """Create a copy of the local sample project, and run simple_deploy
+#     against this project. Most tests will examine how the project
+#     was modified.
+#     """
 
-    # Pause, or the tmpdir won't be usable.
-    sleep(0.2)
+#     # Pause, or the tmpdir won't be usable.
+#     sleep(0.2)
 
-    # Root directory of local simple_deploy project.
-    sd_root_dir = Path(__file__).parent.parent
-    tmp_proj_dir = tmpdir_factory.mktemp('blog_project')
-    cmd = f'sh setup_project.sh -d {tmp_proj_dir} -s {sd_root_dir}'
-    cmd_parts = cmd.split()
-    subprocess.run(cmd_parts)
+#     # Root directory of local simple_deploy project.
+#     sd_root_dir = Path(__file__).parent.parent
+#     tmp_proj_dir = tmpdir_factory.mktemp('blog_project')
+#     cmd = f'sh setup_project.sh -d {tmp_proj_dir} -s {sd_root_dir}'
+#     cmd_parts = cmd.split()
+#     subprocess.run(cmd_parts)
 
-    # Return the location of the temp project.
-    return tmp_proj_dir
+#     # Return the location of the temp project.
+#     return tmp_proj_dir
 
 
 def test_creates_heroku_specific_settings_section(tmp_project):
